+++ inscripta/biocantor/gene/cds.py
@@ -1,6 +1,6 @@
 import warnings
 from itertools import count, zip_longest
-from typing import Iterator, List, Union, Optional, Dict, Hashable, Any, Iterable, Set, Tuple
+from typing import Iterator, List, Union, Optional, Dict, Hashable, Any, Iterable, Set, Tuple, TYPE_CHECKING
 from uuid import UUID
 
 from methodtools import lru_cache
@@ -10,6 +10,7 @@ from inscripta.biocantor.exc import (
     NoSuchAncestorException,
     LocationOverlapException,
     MismatchedFrameException,
+    EmptyLocationException,
 )
 from inscripta.biocantor.gene.cds_frame import CDSPhase, CDSFrame
 from inscripta.biocantor.gene.codon import Codon, TranslationTable
@@ -21,6 +22,9 @@ from inscripta.biocantor.location import Location, Strand, SingleInterval, Compo
 from inscripta.biocantor.parent import Parent, SequenceType
 from inscripta.biocantor.sequence import Sequence, Alphabet
 from inscripta.biocantor.util.hashing import digest_object
+
+if TYPE_CHECKING:
+    from inscripta.biocantor.gene.variants import VariantIntervalCollection, VariantInterval
 
 
 class CDSInterval(AbstractFeatureInterval):
@@ -926,3 +930,24 @@ class CDSInterval(AbstractFeatureInterval):
             InvalidPositionException: If the position provided is not part of this CDSInterval.
         """
         return self.sequence_pos_to_cds(pos) // 3
+
+    def incorporate_variants(self, variants: Union["VariantInterval", "VariantIntervalCollection"]) -> "CDSInterval":
+        """
+        Incorporate all of the variant(s) for an input VariantInterval or VariantIntervalCollection,
+        producing a new CDSInterval with those changes incorporated.
+        """
+        new_loc = variants.lift_over_location(self.chunk_relative_location)
+        if new_loc.is_empty:
+            raise EmptyLocationException("Variant incorporation led to an EmptyLocation")
+        fn = CDSInterval.from_chunk_relative_location if self.is_chunk_relative else CDSInterval.from_location
+        new_frames = CDSInterval.construct_frames_from_location(new_loc, self.frames[0])
+        return fn(
+            new_loc,
+            cds_frames=new_frames,
+            sequence_guid=self.sequence_guid,
+            sequence_name=self.sequence_name,
+            protein_id=self.protein_id,
+            product=self.product,
+            qualifiers=self._export_qualifiers_to_list(),
+            guid=None,  # generate a new Interval GUID based on updated data
+        )